+++ src/components/Container/index.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 
 import './Container.scss';
 
-const Container = ({ children, className }) => (
-  <div className={`container ${className || ''}`}>{children}</div>
+const Container = ({ children, className = '' }) => (
+  <ul className={`container ${className}`}>{children}</ul>
 );
 
 Container.propTypes = {