+++ src/components/Contents/index.js
@@ -1,5 +1,0 @@
-import Song from './Song';
-import Album from './Album';
-import Video from './Video';
-
-export { Song as SongContent, Album as AlbumContent, Video as VideoContent };