+++ src/components/Items/index.js
@@ -1,7 +1,5 @@
 import './Items.scss';
 
-import Album from './Album';
-import Song from './Song';
-import Video from './Video';
-
-export { Album as AlbumItem, Song as SongItem, Video as VideoItem };
+export { default as AlbumItem } from './Album';
+export { default as SongItem } from './Song';
+export { default as VideoItem } from './Video';