+++ src/components/Lists/Artist.js
@@ -8,22 +8,15 @@ import Inline from '../Inline';
 
 import { queryString } from '../../helpers';
 
-const Artist = ({ values, type, className }) => {
+export default function Artist({ values, type, className = '' }) {
   if (values.length > 1) {
     const data = values.map((value, index) => {
       if (value.kind === 'song') {
         return <SongItem key={index} value={value} />;
       } else if (value.collectionType === 'Album') {
-        return <AlbumItem kind="artist" key={index} value={value} />;
+        return <AlbumItem isArtist key={index} value={value} />;
       } else if (value.kind === 'music-video') {
-        return (
-          <VideoItem
-            kind="artist"
-            key={index}
-            value={value}
-            contentList={false}
-          />
-        );
+        return <VideoItem isArtist key={index} value={value} />;
       }
 
       return false;
@@ -43,15 +36,13 @@ const Artist = ({ values, type, className }) => {
           </Link>
         </Inline>
 
-        <div className={`container--horizontal ${className || ''}`}>{data}</div>
+        <ul className={`container--horizontal ${className}`}>{data}</ul>
       </Grid>
     );
   } else return null;
-};
+}
 
 Artist.propTypes = {
   values: PropTypes.array.isRequired,
   type: PropTypes.string.isRequired
 };
-
-export default Artist;