+++ src/components/Row/index.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import './View.scss';
+import './Row.scss';
 
-const View = ({ children, className }) => (
+const Row = ({ children, className = '' }) => (
   <div className={className}>{children}</div>
 );
 
-View.propTypes = {
+Row.propTypes = {
   children: PropTypes.node,
   className: PropTypes.string
 };
 
-export default View;
+export default Row;