+++ src/components/SearchForm/index.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import { withRouter } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
@@ -6,48 +6,38 @@ import { queryString } from '../../helpers';
 
 import './SearchForm.scss';
 
-class SearchForm extends Component {
-  state = { searchTerm: '', visible: false };
+function SearchForm({ history, search }) {
+  const [visible, setVisible] = useState(false);
 
-  handleChange = e => {
+  const handleChange = e => {
     const value = queryString(e.target.value);
 
-    if (value !== '') {
-      this.props.history.push({ pathname: '/search', search: `?q=${value}` });
+    if (value) {
+      history.push({ pathname: '/search', search: `?q=${value}` });
     } else {
-      this.props.history.push({ pathname: '/', search: '' });
+      history.push({ pathname: '/', search: '' });
     }
-
-    this.setState({ searchTerm: e.target.value });
   };
 
-  handleClick = () =>
-    this.state.visible
-      ? this.setState({ visible: false })
-      : this.setState({ visible: true });
+  return (
+    <div className="form__container">
+      <div className={`search__form ${visible ? 'visible' : 'hidden'}`}>
+        <input
+          type="search"
+          placeholder="Search"
+          value={search}
+          onChange={handleChange}
+        />
 
-  render() {
-    return (
-      <div className="form__container">
-        <div
-          className={`search__form ${
-            this.state.visible ? 'visible' : 'hidden'
-          }`}
+        <button
+          className="toggle"
+          onClick={() => (visible ? setVisible(false) : setVisible(true))}
         >
-          <input
-            type="search"
-            placeholder="Search"
-            value={this.props.search}
-            onChange={this.handleChange}
-          />
-
-          <button className="toggle" onClick={this.handleClick}>
-            <FontAwesomeIcon icon="search" />
-          </button>
-        </div>
+          <FontAwesomeIcon icon="search" />
+        </button>
       </div>
-    );
-  }
+    </div>
+  );
 }
 
 export default withRouter(SearchForm);