+++ src/containers/Album/index.js
@@ -1,8 +1,12 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 
 import Preloader from '../../components/Preloader';
-import View from '../../components/View';
-import { AlbumContent } from '../../components/Contents';
+import Col from '../../components/Col';
+import Image from '../../components/Image';
+import ContentHeader from '../../components/ContentHeader';
+import Title from '../../components/Title';
+import { ArtistLink } from '../../components/Links';
+import Info from '../../components/Info';
 import { SongList } from '../../components/SongList';
 import { VideoItem } from '../../components/Items';
 import Grid from '../../components/Grid';
@@ -10,61 +14,93 @@ import Inline from '../../components/Inline';
 
 import './Album.scss';
 
-import { album } from '../../api';
+import { setPageTitle } from '../../helpers';
 
-export default class Album extends Component {
-  state = { album: null, songs: null, videos: null };
+import { getAlbum } from '../../api';
 
-  componentDidMount() {
-    const { id } = this.props.match.params;
+export default function Album(props) {
+  const { id } = props.match.params;
 
-    album(id).then(data => {
-      if (typeof data.results === 'object' && data.results.length > 0) {
-        this.setState({ album: data.results });
+  const [album, setAlbum] = useState(null);
 
-        let songs = [];
-        let videos = [];
+  useEffect(() => {
+    getAlbum(id).then(data => {
+      data = data.results;
 
-        data.results.map((value, index) => {
+      setAlbum(data);
+      setPageTitle(`Album: ${data[0].collectionName}`);
+    });
+  }, []);
+
+  const [songs, setSongs] = useState(null);
+  const [videos, setVideos] = useState(null);
+
+  useEffect(
+    () => {
+      if (album && typeof album === 'object' && album.length > 0) {
+        const songsResults = [];
+        const videosResults = [];
+
+        album.map((value, index) => {
           if (value.kind === 'song') {
-            return songs.push(<SongList key={index} value={value} />);
+            songsResults.push(<SongList key={index} value={value} />);
           } else if (value.kind === 'music-video') {
-            return videos.push(
-              <VideoItem key={index} value={value} contentList={true} />
+            videosResults.push(
+              <VideoItem key={index} value={value} contentList />
             );
           }
 
-          return false;
+          setSongs(songsResults);
+          if (videosResults.length > 0) setVideos(videosResults);
+
+          return null;
         });
-
-        if (videos.length === 0) videos = null;
-
-        this.setState({ songs, videos });
       }
-    });
-  }
 
-  render() {
-    const { album, songs, videos } = this.state;
+      return null;
+    },
+    [album, setSongs, setVideos]
+  );
 
-    return album ? (
-      <View className="album">
-        <AlbumContent value={album[0]}>
-          {songs}
+  return album ? (
+    <article className="album">
+      <Col className="secondary">
+        <aside>
+          <Image
+            className="artwork"
+            src={album[0].artworkUrl100}
+            size="400x400"
+          />
+        </aside>
+      </Col>
 
-          {videos ? (
-            <Grid>
-              <Inline>
-                <h3 className="grid__title">Music videos</h3>
-              </Inline>
+      <Col className="primary">
+        <ContentHeader type="album">
+          <Title
+            title={album[0].collectionName}
+            explicit={album[0].collectionExplicitness}
+          />
 
-              {videos}
-            </Grid>
-          ) : null}
-        </AlbumContent>
-      </View>
-    ) : (
-      <Preloader />
-    );
-  }
+          <ArtistLink value={album[0]} />
+          <Info value={album[0]} />
+        </ContentHeader>
+
+        <ul>{songs}</ul>
+
+        {videos && (
+          <Grid>
+            <Inline>
+              <h3 className="grid__title">Music videos</h3>
+            </Inline>
+
+            <ul>{videos}</ul>
+          </Grid>
+        )}
+
+        <footer className="copyright">{album[0].copyright}</footer>
+      </Col>
+    </article>
+  ) : (
+    <Preloader />
+  );
 }