+++ src/helpers/index.js
@@ -1,5 +1,10 @@
+import json from '../../package.json';
+
 export const queryString = value =>
   value
     .toLowerCase()
     .replace(/[¿@#$%^&/|*?"'`]/g, '')
     .replace(/ /g, '+');
+
+export const setPageTitle = value =>
+  (document.title = `${json.name}${value && ` | ${value}`}`);